pyproject.toml
[project]
name = "llm-examples"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=24.10.0",
    "openai>=1.54.3",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
]
